rig_state_service/requirements.txt
kafka-python==2.0.2
Flask==3.0.2
Flask-Cors==4.0.0